requirements.dev.txt
luma.oled==3.15.0
python-musicpd==0.9.2
ADS1x15-ADC==1.2.2
pyalsaaudio==0.11.0
paho-mqtt==2.1.0
gpiozero==2.0.1
fastapi==0.135.3
uvicorn==0.44.0

#Development
luma.emulator==1.7.0
pygame==2.6.1